+++ src/serde.rs
@@ -6,7 +6,17 @@ impl Serialize for CarettaId {
     where
         S: ::serde::Serializer,
     {
-        serializer.serialize_str(&self.to_string())
+        if serializer.is_human_readable() {
+            let chars = self.to_chars();
+            let mut buf = [0; 7];
+            for i in 0..7 {
+                buf[i] = u8::try_from(chars[i]).unwrap();
+            }
+            let s = unsafe { str::from_utf8_unchecked(&buf) };
+            serializer.serialize_str(s)
+        } else {
+            serializer.serialize_u64(self.to_u64())
+        }
     }
 }
 
@@ -15,7 +25,22 @@ impl<'de> Deserialize<'de> for CarettaId {
     where
         D: ::serde::Deserializer<'de>,
     {
-        let s = String::deserialize(deserializer)?;
-        (&s).parse::<CarettaId>().map_err(|e| D::Error::custom(e))
+        if deserializer.is_human_readable() {
+            #[cfg(feature = "std")]
+            {
+                (&<String as Deserialize>::deserialize(deserializer)?)
+                    .parse::<CarettaId>()
+                    .map_err(D::Error::custom)
+            }
+            #[cfg(not(feature = "std"))]
+            {
+                (<&str as Deserialize>::deserialize(deserializer)?)
+                    .parse::<CarettaId>()
+                    .map_err(D::Error::custom)
+            }
+        } else {
+            let i = u64::deserialize(deserializer)?;
+            CarettaId::from_u64(i).map_err(D::Error::custom)
+        }
     }
 }