+++ tests/rand.rs
@@ -2,6 +2,7 @@
 mod common;
 use caretta_id::*;
 use common::*;
+use rand::Rng;
 
 #[test]
 fn random() {
@@ -15,10 +16,25 @@ fn random() {
 #[test]
 fn random_int() {
     for _ in 0..10 {
-        let value: u64 = rand::random_range(1..<CarettaId>::CAPACITY);
+        let value: u64 = rand::random_range(1..=<CarettaId>::MAX.to_u64());
         let id = <CarettaId>::try_from(value).unwrap();
-        assert_integer_conversion(id);
-        assert_string_convertion(id);
+        assert_conversion(id);
+    }
+}
+
+#[cfg(feature = "std")]
+#[test]
+fn random_str() {
+    let mut rng = rand::rng();
+    for _ in 0..10 {
+        let mut buf = ['0'; 7];
+        for i in 0..6 {
+            let c = rng.sample(rand::distr::Alphanumeric) as char;
+            buf[i] = c;
+        }
+        let s: String = buf.into_iter().collect();
+        let id = (&s).parse::<CarettaId>().unwrap();
+        assert_conversion(id);
     }
 }
 
@@ -26,7 +42,7 @@ fn random_int() {
 #[cfg(feature = "rand")]
 fn oversized_random_int() {
     for _ in 0..10 {
-        let value: u64 = rand::random_range(<CarettaId>::CAPACITY..<u64>::MAX);
+        let value: u64 = rand::random_range((<CarettaId>::MAX.to_u64() + 1)..<u64>::MAX);
         let _ = <CarettaId>::try_from(value).unwrap_err();
     }
 }